Issue-40: android wrappers (#41)
* Changes to enable compilation for Android.

* Set library output path to folder tree with system and processor.

* Moved output of thirdparty dependencies out of build folder.

* Changed channel count to enumeration.

* Added methods for checking if image descriptor is supported.

* Changes for compiling Python wrappers.

* Added error code for case when no physical devices are found.

* Added build_android script to ease flavored ABI builds.

* Update README.md

* Renamed

* Moved to tools dir.

* Final changes before merge.

+++ core/include/lluvia/core/MemoryAllocationInfo.h
@@ -28,7 +28,7 @@ struct MemoryAllocationInfo {
     uint64_t leftPadding;
 
     /** @brief page number within memory */
-    uint64_t page;
+    uint32_t page;
 };
 
 } // namespace ll
